pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiguard"
version = "0.2.0"
description = "AIGuard: model-agnostic safety evaluation toolkit (adversarial, evaluator, hallucination)"
authors = [{ name = "Shelton Mutambirwa", email = "sheltonmutambirwa@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.12.0",
    "pyyaml>=6.0",
]
license = { text = "MIT" }
keywords = ["llm", "safety", "evaluation", "robustness", "hallucination", "adversarial"]
classifiers = [
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/Shelton03/aiguard"
Source = "https://github.com/Shelton03/aiguard"

[project.optional-dependencies]
huggingface = ["datasets>=2.16.0"]
review = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
]
sdk = [
    "litellm>=1.0",
]
monitoring = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
]

[project.scripts]
aiguard = "aiguard.cli.main:app"
aiguard-review = "review.cli:main"

[tool.setuptools]
packages = [
    "aiguard",
    "aiguard.cli",
    "aiguard.evaluation",
    "aiguard.sdk",
    "adversarial",
    "adversarial.adapters",
    "evaluator",
    "hallucination",
    "storage",
    "review",
    "pipeline",
    "monitoring",
    "monitoring.api",
    "monitoring.services",
    "config",
]
include-package-data = true

[tool.setuptools.package-data]
"adversarial" = ["py.typed"]
"evaluator"   = ["py.typed"]
"hallucination" = ["py.typed"]
"review"      = ["templates/*.html", "static/*.css"]
